Refactor the Card class and add lazy loading to the highestAttribute calculation
